refactor: 하나의 모달에서 처리
